This is a command line connect four game made with ruby.
This was my first project using a TDD approach,
and while I attempted to write tests first, I did not
strictly stay on the TDD path, sometimes wandering away
to write something and then going back to write the tests.

My ability to anticipate the correct expectations for a method
is developing, and I can't always predict what needs to be tested
before writing some code. I think that will come with time.

The game itself was easy to make, being very similar to
tic-tac-toe.